Factor out HTTP client for API.
